+++ src/homeii-music-flow.js
@@ -93,9 +93,9 @@ function ensureHaEditorComponents() {
   } catch (_) {}
 }
 
-const HOMEII_CARD_VERSION = "5.8.2-beta.2";
-const HOMEII_BROWSER_EDITOR_TAG = "homeii-music-flow-browser-editor-v5822";
-const HOMEII_MOBILE_EDITOR_TAG = "homeii-music-flow-editor-v5822";
+const HOMEII_CARD_VERSION = "5.8.2-beta.3";
+const HOMEII_BROWSER_EDITOR_TAG = "homeii-music-flow-browser-editor-v5823";
+const HOMEII_MOBILE_EDITOR_TAG = "homeii-music-flow-editor-v5823";
 const AMBIENT_LIGHT_PAIR_PLAYER_PREFIX = "__homeii_ambient_light_pair_player_";
 const AMBIENT_LIGHT_PAIR_LIGHTS_PREFIX = "__homeii_ambient_light_pair_lights_";